move device detection into separate files
clean up sicm_init
