+++ src/low/detect_devices.c
@@ -1,0 +1,44 @@
+#include <numa.h>
+
+#include "detect_devices.h"
+
+static struct bitmask *get_compute_nodes(int node_count) {
+    struct bitmask* compute_nodes = numa_bitmask_alloc(node_count);
+    struct bitmask* cpumask = numa_allocate_cpumask();
+    int cpu_count = numa_num_possible_cpus();
+    for(int i = 0; i < node_count; i++) {
+        numa_node_to_cpus(i, cpumask);
+        for(int j = 0; j < cpu_count; j++) {
+            if(numa_bitmask_isbitset(cpumask, j)) {
+                numa_bitmask_setbit(compute_nodes, i);
+                break;
+            }
+        }
+    }
+    numa_free_cpumask(cpumask);
+    return compute_nodes;
+}
+
+int detect_devices(int node_count,
+                   int *huge_page_sizes, int huge_page_size_count, int normal_page_size,
+                   struct sicm_device **devices) {
+    int idx = 0;
+
+    struct bitmask* compute_nodes = get_compute_nodes(node_count);
+
+    struct bitmask* non_dram_nodes = numa_bitmask_alloc(node_count);
+    detect_x86    (compute_nodes, non_dram_nodes,
+                   huge_page_sizes, huge_page_size_count, normal_page_size,
+                   devices, &idx);
+
+    detect_powerpc(compute_nodes, non_dram_nodes,
+                   huge_page_sizes, huge_page_size_count, normal_page_size,
+                   devices, &idx);
+
+    detect_DRAM   (compute_nodes, non_dram_nodes,
+                   huge_page_sizes, huge_page_size_count, normal_page_size,
+                   devices, &idx);
+    numa_bitmask_free(non_dram_nodes);
+    numa_bitmask_free(compute_nodes);
+    return idx;
+}
