+++ src/low/sicm_low.c
@@ -18,9 +18,7 @@
 #include <linux/mman.h>
 #endif
 #include "sicm_impl.h"
-
-#define X86_CPUID_MODEL_MASK        (0xf<<4)
-#define X86_CPUID_EXT_MODEL_MASK    (0xf<<16)
+#include "detect_devices.h"
 
 int normal_page_size = -1;
 
@@ -87,29 +85,23 @@ struct sicm_device_list sicm_init() {
 
   // Find the number of huge page sizes
   int huge_page_size_count = 0;
-  DIR* dir;
-  struct dirent* entry;
-  dir = opendir("/sys/kernel/mm/hugepages");
+
+  DIR* dir = opendir("/sys/kernel/mm/hugepages");
+  struct dirent* entry = NULL;
   while((entry = readdir(dir)) != NULL)
     if(entry->d_name[0] != '.') huge_page_size_count++;
-  closedir(dir);
 
-  int node_count = numa_max_node() + 1, depth;
-  int device_count = node_count * (huge_page_size_count + 1);
-
-  struct bitmask* non_dram_nodes = numa_bitmask_alloc(node_count);
+  const int node_count = numa_max_node() + 1, depth;
+  const int device_count = node_count * (huge_page_size_count + 1);
 
   sicm_global_device_array = malloc(device_count * sizeof(struct sicm_device));
   int* huge_page_sizes = malloc(huge_page_size_count * sizeof(int));
-
-  int i, j;
-  int idx = 0;
 
   normal_page_size = numa_pagesize() / 1024;
 
   // initialize the device list
   sicm_device **devices = malloc(device_count * sizeof(sicm_device *));
-  for(i = 0; i < device_count; i++) {
+  for(int i = 0; i < device_count; i++) {
       devices[i] = &sicm_global_device_array[i];
       devices[i]->tag = INVALID_TAG;
       devices[i]->node = -1;
@@ -117,11 +109,12 @@ struct sicm_device_list sicm_init() {
   }
 
   // Find the actual set of huge page sizes (reported in KiB)
-  dir = opendir("/sys/kernel/mm/hugepages");
-  i = 0;
+  rewinddir(dir);
+  int i = 0;
   while((entry = readdir(dir)) != NULL) {
     if(entry->d_name[0] != '.') {
       huge_page_sizes[i] = 0;
+      int j;
       for(j = 0; j < 10; j++) {
         if(entry->d_name[j] == '\0') {
           j = -1;
@@ -138,155 +131,11 @@ struct sicm_device_list sicm_init() {
   }
   closedir(dir);
 
-  struct bitmask* cpumask = numa_allocate_cpumask();
-  int cpu_count = numa_num_possible_cpus();
-  struct bitmask* compute_nodes = numa_bitmask_alloc(node_count);
-  i = 0;
-  for(i = 0; i < node_count; i++) {
-    numa_node_to_cpus(i, cpumask);
-    for(j = 0; j < cpu_count; j++) {
-      if(numa_bitmask_isbitset(cpumask, j)) {
-        numa_bitmask_setbit(compute_nodes, i);
-        break;
-      }
-    }
-  }
-  numa_free_cpumask(cpumask);
+  const int idx = detect_devices(node_count,
+                                 huge_page_sizes, huge_page_size_count,
+                                 normal_page_size,
+                                 devices);
 
-  #ifdef __x86_64__
-  // Knights Landing
-  uint32_t xeon_phi_model = (0x7<<4);
-  uint32_t xeon_phi_ext_model = (0x5<<16);
-  uint32_t registers[4];
-  uint32_t expected = xeon_phi_model | xeon_phi_ext_model;
-  asm volatile("cpuid":"=a"(registers[0]),
-                       "=b"(registers[1]),
-                       "=c"(registers[2]),
-                       "=d"(registers[2]):"0"(1), "2"(0));
-  uint32_t actual = registers[0] & (X86_CPUID_MODEL_MASK | X86_CPUID_EXT_MODEL_MASK);
-
-  if (actual == expected) {
-    for(i = 0; i <= numa_max_node(); i++) {
-      if(!numa_bitmask_isbitset(compute_nodes, i)) {
-        long size = -1;
-        if ((numa_node_size(i, &size) != -1) && size) {
-          int compute_node = -1;
-          /*
-           * On Knights Landing machines, high-bandwidth memory always has
-           * higher NUMA distances (to prevent malloc from giving you HBM)
-           * but I'm pretty sure the compute node closest to an HBM node
-           * always has NUMA distance 31, e.g.,
-           * https://goparallel.sourceforge.net/wp-content/uploads/2016/05/Colfax_KNL_Clustering_Modes_Guide.pdf
-           */
-          for(j = 0; j < numa_max_node(); j++) {
-              if(numa_distance(i, j) == 31) compute_node = j;
-          }
-          devices[idx]->tag = SICM_KNL_HBM;
-          devices[idx]->node = i;
-          devices[idx]->page_size = normal_page_size;
-          devices[idx]->data.knl_hbm = (struct sicm_knl_hbm_data){
-            .compute_node=compute_node };
-          numa_bitmask_setbit(non_dram_nodes, i);
-          idx++;
-          for(j = 0; j < huge_page_size_count; j++) {
-              devices[idx]->tag = SICM_KNL_HBM;
-              devices[idx]->node = i;
-              devices[idx]->page_size = huge_page_sizes[j];
-              devices[idx]->data.knl_hbm = (struct sicm_knl_hbm_data){
-                .compute_node=compute_node };
-              idx++;
-          }
-        }
-      }
-    }
- } else {
-   // Optane support
-   // This is a bit of a hack: on x86_64 architecture that is not KNL,
-   // NUMA nodes without CPUs are assumed to be Optane nodes
-   for(i = 0; i <= numa_max_node(); i++) {
-     if(!numa_bitmask_isbitset(compute_nodes, i)) {
-       long size = -1;
-       if ((numa_node_size(i, &size) != -1) && size) {
-         int compute_node = -1;
-	 int dist = 1000;
-         for(j = 0; j < numa_max_node(); j++) {
-	   if (i == j)
-	     continue;
-	   int d = numa_distance(i, j);
-	   if (d < dist) {
-		dist = d;
-		compute_node = j;
-	   }
-         }
-         devices[idx]->tag = SICM_OPTANE;
-         devices[idx]->node = i;
-         devices[idx]->page_size = normal_page_size;
-         devices[idx]->data.optane = (struct sicm_optane_data){
-           .compute_node=compute_node };
-         numa_bitmask_setbit(non_dram_nodes, i);
-         idx++;
-         for(j = 0; j < huge_page_size_count; j++) {
-             devices[idx]->tag = SICM_OPTANE;
-             devices[idx]->node = i;
-             devices[idx]->page_size = huge_page_sizes[j];
-             devices[idx]->data.optane = (struct sicm_optane_data){
-               .compute_node=compute_node };
-             idx++;
-         }
-       }
-     }
-   }
-  }
-  #endif
-
-  #ifdef __powerpc__
-  // Power PC
-  for(i = 0; i <= numa_max_node(); i++) {
-    if(!numa_bitmask_isbitset(compute_nodes, i)) {
-      // make sure the numa node has memory on it
-      long size = -1;
-      if ((numa_node_size(i, &size) != -1) && size) {
-        devices[idx]->tag = SICM_POWERPC_HBM;
-        devices[idx]->node = i;
-        devices[idx]->page_size = normal_page_size;
-        devices[idx]->data.powerpc_hbm = (struct sicm_powerpc_hbm_data){ };
-        numa_bitmask_setbit(non_dram_nodes, i);
-        idx++;
-        for(j = 0; j < huge_page_size_count; j++) {
-          devices[idx]->tag = SICM_POWERPC_HBM;
-          devices[idx]->node = i;
-          devices[idx]->page_size = huge_page_sizes[j];
-          devices[idx]->data.powerpc_hbm = (struct sicm_powerpc_hbm_data){ };
-          idx++;
-        }
-      }
-    }
-  }
-  #endif
-
-  // DRAM
-  for(i = 0; i <= numa_max_node(); i++) {
-    if(!numa_bitmask_isbitset(non_dram_nodes, i)) {
-      long size = -1;
-      if ((numa_node_size(i, &size) != -1) && size) {
-        devices[idx]->tag = SICM_DRAM;
-        devices[idx]->node = i;
-        devices[idx]->page_size = normal_page_size;
-        devices[idx]->data.dram = (struct sicm_dram_data){ };
-        idx++;
-        for(j = 0; j < huge_page_size_count; j++) {
-          devices[idx]->tag = SICM_DRAM;
-          devices[idx]->node = i;
-          devices[idx]->page_size = huge_page_sizes[j];
-          devices[idx]->data.dram = (struct sicm_dram_data){ };
-          idx++;
-        }
-      }
-    }
-  }
-
-  numa_bitmask_free(compute_nodes);
-  numa_bitmask_free(non_dram_nodes);
   free(huge_page_sizes);
 
   qsort(devices, idx, sizeof(sicm_device *), sicm_device_compare);
